refactor: moving remotesecretfilter into its own file
