+++ pkg/serviceprovider/tokenfilter_test.go
@@ -19,10 +19,6 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/redhat-appstudio/remote-secret/api/v1beta1"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
 	api "github.com/redhat-appstudio/service-provider-integration-operator/api/v1beta1"
 	"github.com/stretchr/testify/assert"
 )
@@ -71,61 +67,4 @@ func TestTokenFilterFunc_Matches(t *testing.T) {
 			assert.Equalf(t, tt.want, got, "Matches(%v, %v, %v)", tt.args.ctx, tt.args.matchable, tt.args.token)
 		})
 	}
-}
-
-func TestDefaultRemoteSecretFilterFunc(t *testing.T) {
-	remoteSecret := v1beta1.RemoteSecret{
-		TypeMeta: metav1.TypeMeta{},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "rs",
-			Namespace: "ns",
-		},
-		Spec: v1beta1.RemoteSecretSpec{
-			Secret: v1beta1.LinkableSecretSpec{
-				Name: "secret",
-				Type: corev1.SecretTypeBasicAuth,
-			},
-			Targets: []v1beta1.RemoteSecretTarget{{
-				Namespace: "ns",
-			}},
-		},
-		Status: v1beta1.RemoteSecretStatus{
-			Conditions: []metav1.Condition{{
-				Type:   string(v1beta1.RemoteSecretConditionTypeDataObtained),
-				Status: metav1.ConditionTrue,
-			}},
-			Targets: []v1beta1.TargetStatus{{
-				Namespace:  "ns",
-				SecretName: "secret",
-			}},
-		},
-	}
-	accessCheck := api.SPIAccessCheck{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "accessCheck",
-			Namespace: "ns",
-		},
-	}
-
-	t.Run("all conditions satisfied", func(t *testing.T) {
-		assert.True(t, DefaultRemoteSecretFilterFunc(context.TODO(), &accessCheck, &remoteSecret))
-	})
-
-	t.Run("data not obtained", func(t *testing.T) {
-		rs := remoteSecret.DeepCopy()
-		rs.Status.Conditions[0].Status = metav1.ConditionFalse
-		assert.False(t, DefaultRemoteSecretFilterFunc(context.TODO(), &accessCheck, rs))
-	})
-
-	t.Run("wrong secret type", func(t *testing.T) {
-		rs := remoteSecret.DeepCopy()
-		rs.Spec.Secret.Type = corev1.SecretTypeOpaque
-		assert.False(t, DefaultRemoteSecretFilterFunc(context.TODO(), &accessCheck, rs))
-	})
-
-	t.Run("target in different namespace", func(t *testing.T) {
-		rs := remoteSecret.DeepCopy()
-		rs.Status.Targets[0].Namespace = "diff-ns"
-		assert.False(t, DefaultRemoteSecretFilterFunc(context.TODO(), &accessCheck, rs))
-	})
 }
